feat: replace all places where global constants are referred with references to newly generated GlobalConstants class (#714)
Resolves: #711

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/limits/precompiles/Blake2fRounds.java
@@ -44,7 +44,7 @@ public final class Blake2fRounds implements Module {
 
   @Getter private final Blake2fModexpData blake2fModexpData;
 
-  private int lastDataCallHubStamp = 0;
+  private long lastDataCallHubStamp = 0;
 
   private final Stack<Integer> counts = new Stack<>();
